Address data races and timing issues in TLSCredProcessor and its tests
Summary:
The two callback vectors in TLSCredProcessor are written and read by different threads. There was no synchronization between them so I added some through `folly::Synchronized`. Writers lock to add, readers get a copy to iterate.

I explained in T98770325 how the timing of the FilePoller interacted with the file changes (write + update time) and the `Baton`. This diff changes the tests to wait after each file modification so there's no doubt about why the callback was triggered.

TSAN complained about all the booleans being read from and written to by different threads. I think that aspect of the code was safe considering the batons, but they didn't seem necessary so I removed them. The fact that the batons are set/unset is signal enough.

Reviewed By: samuelrm

Differential Revision: D30945100

fbshipit-source-id: 05e84a5ef7ba3a8e53ad2a78ee306929b7ab3f4b