SSLContext: rename misleading loadClientCAList
Summary:
`loadClientCAList`, unlike `loadTrustedCertificates`, does not *append*
"client ca"s to the current context, it *overrides* the list of acceptable
client CA names.

`load` vs `set` have subtle, but recognizable, semantic differences. There
is a currently a bug in wangle::SSLContextManager that could have been
prevented if the name was more wisely chosen.

Rename to `setSupportedClientCertificateAuthorityNames`, which is verbose but
* Makes the `load` vs `set` distinction clear.
* Makes it clear that this has nothing to do with changing how client CAs are
  configured. All this does is set a list of supported names; nothing more.
  Existing call sites imply that authors misunderstood what this function
  was doing.

Reviewed By: kylemirz

Differential Revision: D32408104

fbshipit-source-id: acf55c2e0b206e28e13e5922d719b43289a1c0f7