Fix a race in AcceptRoutingHandler
Summary:
This is the same issue and the same fix as in D19873260. Description from that diff applies to this issue as well:
It's not safe to pass a socket to another thread before its DelayedDestruction destructor guard count becomes 0 (it's always non-0 if detach happens from within one of the socket callbacks).

Reviewed By: yfeldblum

Differential Revision: D34381360

fbshipit-source-id: 2aa5cc1a2a6e0ad92c25c00f4b64174cf6e90e4f