Extend ServerBootstap to accept external connections
Summary:
Right now ServerBootstrap is optimized for managing all the accepting sockets directly. This makes it harder to dynamically start and stop listening on some sockets.

This diff introduces a mechanism to add connections to ServerBootstrap even when they were accepted externally. The new API also enables passing an observer that has to be attached to the socket, since Acceptor currently encapsulates socket creation.

Reviewed By: jkedgar

Differential Revision: D36383876

fbshipit-source-id: ef7c52a872f104274817b30fe0ad9575b249e7da