Update addSSLContextConfig to not try to load empty client ca files
Summary:
D37429044 caused a test to fail due to it setting an empty client_ca_file (specifically testSSLClient and testTickets in thrift/facebook/test/py:cpp_server_test), the tests would timeout due to the ca file being passed in being empty (client_ca_path is set to "" here https://fburl.com/code/tnc9ur4b).

To reach parity, we would need to make sure that no empty ca files get loaded in addSSLContextConfig, alternatively if we want to fail fast when client ca files are empty we would need to update the test. The latter makes sense to me, but I assume that there was a reason for adding the !empty check to the non-vector clientCAFile in addSSLContextConfig?

Reviewed By: xybu

Differential Revision: D38264435

fbshipit-source-id: 94cd98ba4bd5f7c49f47b2c40feb581d44b19d3e