Change resetTimeoutTo to schedule the timeout even if the specified timeout is zero
Summary:
scheduleTimeout/resetTimeout have an unintuitive behavior of NOT scheduling the timeout at all if the value is 0.  Presumably this was to easily implement 0 == disabled idle timeout for ConnectionManager.

However, there's a case where we want to resetTimeoutTo(0) and have the timeout fire, so refactor to make that possible, and document the subtle behavior

Reviewed By: mjoras

Differential Revision: D62394198

fbshipit-source-id: bcd93ff12f4bf5784dd9cd6cf7a94b5473929c2e