TransportPeeker: check for io_uring zc rx support
Summary:
TransportPeeker is used for peeking sockets for crypto and Thrift. Check whether the underlying EVB backend for the AsyncTransport to be peeked supports io_uring zc rx. And if so set the ReadMode to ReadRecvZc such that the AsyncSocket will issue a recvzc read.

# Context
This and the rest of the stack is needed to support peeking done at the crypto layer in wangle::Acceptor to determine things like StopTLS and at the Thrift layer to determine the protocol. 9 bytes are peeked for crypto and 13 bytes are peeked for Thrift.

The peeking happens early on before an AsyncIoUringSocket is created on top of AsyncSSLSocket. Prior to this, the only way to read a socket configured for io_uring zc rx is via AsyncIoUringSocket.

Reviewed By: dmm-fb

Differential Revision: D70057957

fbshipit-source-id: 0ef2ed246167ac7d1645512e5585d78af4ba3878